+++ tests/testthat/test_gld.R
@@ -6,7 +6,8 @@ test_that(desc = 'Test of main function',{
   f_1 <- system.file("extdata/3262699324398819.zip", package = 'GetLattesData')
   f_2 <- system.file("extdata/8373564643000623.zip", package = 'GetLattesData')
 
-  l_lattes <- gld_get_lattes_data_from_zip(c(f_1, f_2))
+  l_lattes <- gld_get_lattes_data_from_zip(c(f_1, f_2),
+                                           field.qualis = 'ADMINISTRAÇÃO PÚBLICA E DE EMPRESAS, CIÊNCIAS CONTÁBEIS E TURISMO')
 
   expect_true(is.list(l_lattes))
   expect_true(length(l_lattes) > 0)